full model flow added, seems okay, water temp to be added

+++ 2.code/0.run.full.r
@@ -24,14 +24,14 @@ data.formatted <- 0 # 0 no, 1 yes
 autocor.incl <- 1 # 0 for no, 1 for yes
 
 # mcmc param first model
-mcmc.burn.1 <- as.integer(2000)
-mcmc.length.1 <- as.integer(4000)  
+mcmc.burn.1 <- as.integer(1000)
+mcmc.length.1 <- as.integer(2000)  
 mcmc.thin.1 = 5
 mcmc.chains.1 = 3 # needs
 
 # mcmc param first model
-mcmc.burn.2 <- as.integer(3000)
-mcmc.length.2 <- as.integer(6000)  
+mcmc.burn.2 <- as.integer(1000)
+mcmc.length.2 <- as.integer(2000)  
 mcmc.thin.2 = 5
 mcmc.chains.2 = 3 # needs
